DDCNL-8950: Moved total tax calculation methods to helper
