+++ app/uk/gov/hmrc/tai/service/TotalTaxService.scala
@@ -17,35 +17,34 @@
 package uk.gov.hmrc.tai.service
 
 import com.google.inject.{Inject, Singleton}
-import play.api.libs.json.JsValue
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.tai.connectors.TaxAccountConnector
-import uk.gov.hmrc.tai.model.domain._
-import uk.gov.hmrc.tai.model.domain.calculation.{CodingComponent, IncomeCategory, TotalTax}
-import uk.gov.hmrc.tai.model.domain.formatters.taxComponents.TaxAccountHodFormatters
+import uk.gov.hmrc.tai.model.domain.calculation.{IncomeCategory, TotalTax}
 import uk.gov.hmrc.tai.model.domain.formatters.{IncomeCategoryHodFormatters, TaxAccountSummaryHodFormatters}
-import uk.gov.hmrc.tai.model.domain.taxAdjustments.{AlreadyTaxedAtSource, OtherTaxDue, ReliefsGivingBackTax, TaxAdjustment, TaxAdjustmentComponent, TaxReliefComponent}
 import uk.gov.hmrc.tai.model.tai.TaxYear
+import uk.gov.hmrc.tai.service.helper.TaxAccountHelper
 
 import scala.concurrent.{ExecutionContext, Future}
 
 @Singleton
 class TotalTaxService @Inject() (
-  taxAccountConnector: TaxAccountConnector
+  taxAccountConnector: TaxAccountConnector,
+  taxAccountHelper: TaxAccountHelper,
+  taxAccountSummaryService: TaxAccountSummaryService
 )(implicit ec: ExecutionContext)
-    extends IncomeCategoryHodFormatters with TaxAccountSummaryHodFormatters with TaxAccountHodFormatters {
+    extends IncomeCategoryHodFormatters with TaxAccountSummaryHodFormatters {
 
   def totalTax(nino: Nino, year: TaxYear)(implicit hc: HeaderCarrier): Future[TotalTax] = {
     val taxAccountDetails = taxAccountConnector.taxAccount(nino, year)
     for {
       incomeCategories     <- taxAccountDetails.map(_.as[Seq[IncomeCategory]](incomeCategorySeqReads))
-      totalTaxAmount       <- taxAccountSummary(taxAccountDetails)
-      reliefsGivingBackTax <- reliefsGivingBackTaxComponents(taxAccountDetails)
-      otherTaxDue          <- otherTaxDueComponents(taxAccountDetails)
-      alreadyTaxedAtSource <- alreadyTaxedAtSourceComponents(taxAccountDetails)
+      totalTaxAmount       <- taxAccountSummaryService.totalTaxAmount(nino, year)
+      reliefsGivingBackTax <- taxAccountHelper.reliefsGivingBackTaxComponents(taxAccountDetails)
+      otherTaxDue          <- taxAccountHelper.otherTaxDueComponents(taxAccountDetails)
+      alreadyTaxedAtSource <- taxAccountHelper.alreadyTaxedAtSourceComponents(taxAccountDetails)
       taxOnOtherIncome     <- taxAccountDetails.map(_.as[Option[BigDecimal]](taxOnOtherIncomeRead))
-      taxReliefComponents  <- taxReliefComponents(taxAccountDetails)
+      taxReliefComponents  <- taxAccountHelper.taxReliefComponents(taxAccountDetails)
     } yield TotalTax(
       totalTaxAmount,
       incomeCategories,
@@ -55,132 +54,6 @@ class TotalTaxService @Inject() (
       taxOnOtherIncome,
       taxReliefComponents
     )
-  }
-
-  def taxAccountSummary(taxAccountDetails: Future[JsValue]): Future[BigDecimal] = {
-    val componentTypesCanAffectTotalEst: Seq[TaxComponentType] =
-      Seq(UnderPaymentFromPreviousYear, OutstandingDebt, EstimatedTaxYouOweThisYear)
-
-    taxAccountDetails.flatMap { taxAccount =>
-      val totalTax = taxAccount.as[BigDecimal](taxAccountSummaryReads)
-      val componentsCanAffectTotal = taxAccount
-        .as[Seq[CodingComponent]](codingComponentReads)
-        .filter(c => componentTypesCanAffectTotalEst.contains(c.componentType))
-      Future(totalTax + componentsCanAffectTotal.map(_.inputAmount.getOrElse(BigDecimal(0))).sum)
-    }
-  }
-
-  def reliefsGivingBackTaxComponents(taxAccountDetails: Future[JsValue]): Future[Option[TaxAdjustment]] = {
-    val reliefsGivingBackTaxComponents = taxAdjustmentComponents(taxAccountDetails).map {
-      case Some(taxAdjustment) =>
-        taxAdjustment.taxAdjustmentComponents.filter {
-          _.taxAdjustmentType match {
-            case _: ReliefsGivingBackTax => true
-            case _                       => false
-          }
-        }
-      case None => Seq.empty[TaxAdjustmentComponent]
-    }
-
-    for {
-      reliefComponents <- reliefsGivingBackTaxComponents
-    } yield
-      if (reliefComponents.nonEmpty) {
-        Some(TaxAdjustment(reliefComponents.map(_.taxAdjustmentAmount).sum, reliefComponents))
-      } else {
-        None
-      }
-  }
-
-  def otherTaxDueComponents(taxAccountDetails: Future[JsValue]): Future[Option[TaxAdjustment]] = {
-    val otherTaxDueComponents = taxAdjustmentComponents(taxAccountDetails).map {
-      case Some(taxAdjustment) =>
-        taxAdjustment.taxAdjustmentComponents.filter {
-          _.taxAdjustmentType match {
-            case _: OtherTaxDue => true
-            case _              => false
-          }
-        }
-      case None => Seq.empty[TaxAdjustmentComponent]
-    }
-
-    for {
-      otherTaxComponents <- otherTaxDueComponents
-    } yield
-      if (otherTaxComponents.nonEmpty) {
-        Some(TaxAdjustment(otherTaxComponents.map(_.taxAdjustmentAmount).sum, otherTaxComponents))
-      } else {
-        None
-      }
-  }
-
-  def alreadyTaxedAtSourceComponents(taxAccountDetails: Future[JsValue]): Future[Option[TaxAdjustment]] = {
-    val alreadyTaxedSourcesComponents = taxAdjustmentComponents(taxAccountDetails).map {
-      case Some(taxAdjustment) =>
-        taxAdjustment.taxAdjustmentComponents.filter {
-          _.taxAdjustmentType match {
-            case _: AlreadyTaxedAtSource => true
-            case _                       => false
-          }
-        }
-      case None => Seq.empty[TaxAdjustmentComponent]
-    }
-
-    for {
-      alreadyTaxedAtSourceComponents <- alreadyTaxedSourcesComponents
-    } yield
-      if (alreadyTaxedAtSourceComponents.nonEmpty) {
-        Some(
-          TaxAdjustment(alreadyTaxedAtSourceComponents.map(_.taxAdjustmentAmount).sum, alreadyTaxedAtSourceComponents)
-        )
-      } else {
-        None
-      }
-  }
-
-  def taxReliefComponents(taxAccountDetails: Future[JsValue]): Future[Option[TaxAdjustment]] = {
-    lazy val taxReliefsComponentsFuture = taxAdjustmentComponents(taxAccountDetails).map {
-      case Some(taxAdjustment) =>
-        taxAdjustment.taxAdjustmentComponents.filter {
-          _.taxAdjustmentType match {
-            case _: TaxReliefComponent => true
-            case _                     => false
-          }
-        }
-      case None => Seq.empty[TaxAdjustmentComponent]
-    }
-
-    lazy val codingComponentFuture = taxAccountDetails.map(
-      _.as[Seq[CodingComponent]](codingComponentReads)
-    )
-
-    for {
-      taxReliefComponents <- taxReliefsComponentsFuture
-      codingComponents    <- codingComponentFuture
-    } yield {
-      val giftAidPayments = codingComponents.find(_.componentType == GiftAidPayments).flatMap(_.inputAmount)
-      val components = giftAidPayments.collect { case amount =>
-        taxReliefComponents :+ TaxAdjustmentComponent(taxAdjustments.GiftAidPayments, amount)
-      } getOrElse taxReliefComponents
-
-      if (components.nonEmpty) Some(TaxAdjustment(components.map(_.taxAdjustmentAmount).sum, components)) else None
-    }
-  }
-
-  def taxAdjustmentComponents(taxAccountDetails: Future[JsValue]): Future[Option[TaxAdjustment]] = {
-
-    val taxAdjustmentComponents = taxAccountDetails.map(
-      _.as[Seq[TaxAdjustmentComponent]](taxAdjustmentComponentReads)
-    )
-
-    for {
-      taxAdjustments <- taxAdjustmentComponents
-    } yield
-      if (taxAdjustments.nonEmpty) {
-        Some(TaxAdjustment(taxAdjustments.map(_.taxAdjustmentAmount).sum, taxAdjustments))
-      } else {
-        None
-      }
   }
 
   def taxFreeAllowance(nino: Nino, year: TaxYear)(implicit hc: HeaderCarrier): Future[BigDecimal] =
